PacketUploadService and test cases moved to app
