+++ client/app/build.gradle
@@ -82,6 +82,9 @@ dependencies {
     testImplementation "junit:junit:4.13.2"
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+    androidTestImplementation 'com.squareup.okhttp3:mockwebserver:3.14.9'
+    androidTestImplementation "org.mockito:mockito-android:4.6.0"
+    androidTestImplementation "androidx.room:room-runtime:2.4.2"
 
     // (Java only)
     implementation "androidx.work:work-runtime:2.7.1"
