+++ client/clientmanager/src/main/java/io/mosip/registration/clientmanager/config/AppModule.java
@@ -15,8 +15,6 @@ import io.mosip.registration.clientmanager.service.*;
 import io.mosip.registration.clientmanager.BuildConfig;
 import io.mosip.registration.clientmanager.service.JobManagerServiceImpl;
 import io.mosip.registration.clientmanager.service.JobTransactionServiceImpl;
-import io.mosip.registration.clientmanager.service.PacketUploadServiceImpl;
-import io.mosip.registration.clientmanager.spi.PacketUploadService;
 import io.mosip.registration.clientmanager.util.DateUtil;
 import io.mosip.registration.clientmanager.spi.JobManagerService;
 import io.mosip.registration.clientmanager.spi.AuditManagerService;
@@ -212,12 +210,6 @@ public class AppModule {
     @Singleton
     DateUtil provideDateUtil() {
         return new DateUtil(appContext);
-    }
-
-    @Provides
-    @Singleton
-    PacketUploadService providePacketUploadService(PacketService packetService) {
-        return new PacketUploadServiceImpl(packetService);
     }
 
     @Provides
